requirements.txt
streamlit
Pillow
streamlit_option_menu
numpy